% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_info.R
\name{add_group_info}
\alias{add_group_info}
\title{Add grouping information of the WDI data to a metric summary}
\usage{
add_group_info(metric_summary, wdi_data)
}
\arguments{
\item{metric_summary}{A data frame containing the calculated diagnostic indices generated by any of the following functions:
\code{\link{compute_variation}}, \code{\link{compute_trend_shape_features}}, \code{\link{compute_temporal_features}}, or \code{\link{compute_diagnostic_indices}}}

\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}
}
\value{
A data frame containing the calculated diagnostic indices and the grouping variables in the WDI data set.
}
\description{
Add grouping information of the WDI data to a metric summary
}
\examples{
pm_diagnostic_metrics <- compute_diagnostic_indices(pm_data, group_var = "region")
pm_diagnostic_metrics_group <- add_group_info(metric_summary = pm_diagnostic_metrics,pm_data)
}
