% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_wdi_data.R
\name{cache_wdi_data}
\alias{cache_wdi_data}
\title{Cache WDI data of the specified indicator}
\usage{
cache_wdi_data(indicator)
}
\arguments{
\item{indicator}{any data indicator from the WDI website}
}
\value{
a .rds containing the data  set of the specified indicator
}
\description{
Cache WDI data of the specified indicator
}
\examples{
data <- cache_wdi_data(indicator = "LO.PISA.MAT")
}
