% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dissimilarity.R
\name{compute_dissimilarity}
\alias{compute_dissimilarity}
\title{compute daisy dissimilarity distance to WDI data
compute dissimilarity distance and convert the output to matrix}
\usage{
compute_dissimilarity(data, index, metric = "euclidean")
}
\arguments{
\item{data}{A tibble either generated by \code{\link{cache_wdi_data}} or downloaded from other world bank-hosted databases}

\item{index}{Any indicator of choice or the name of the variable containing the observational data}

\item{metric}{A default euclidean metric, if otherwise, specify the type of method to calculate dissimilarity distance}
}
\value{
A matrix indicating the pairwise comparisons
}
\description{
compute daisy dissimilarity distance to WDI data
compute dissimilarity distance and convert the output to matrix
}
\examples{
dist_mat <- compute_dissimilarity(data = pisa_data, index = "LO.PISA.MAT")
}
\seealso{
\code{\link[=cache_wdi_data]{cache_wdi_data()}} to load/cache the data
}
\author{
Oluwayomi Akinfenwa - \if{html}{\out{<\email{yomiolaitan03@gmail.com}>}}
}
