% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_smoothness.R
\name{compute_smoothness}
\alias{compute_smoothness}
\title{Compute smoothness measures to the WDI data or data from other world bank-hosted databases}
\usage{
compute_smoothness(data, index)
}
\arguments{
\item{data}{A tibble either generated by \code{\link{cache_wdi_data}} or downloaded from other world bank-hosted databases}

\item{index}{Any indicator of choice or the name of the variable containing the observational data}
}
\value{
A data frame with the smoothness measures
}
\description{
Compute smoothness measures to the WDI data or data from other world bank-hosted databases
}
\examples{

smoothness <- compute_smoothness(data = pisa_data, index = "LO.PISA.MAT")
}
