% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_temporal_features.R
\name{compute_temporal_features}
\alias{compute_temporal_features}
\title{Compute sequential temporal features}
\usage{
compute_temporal_features(wdi_data, index = NULL)
}
\arguments{
\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}

\item{index}{An optional character string specifying the indicator code
Defaults to \code{NULL}}
}
\value{
A data frame with columns \code{country}, \code{crossing_points}, \code{flat_spot}, and \code{acf}.
}
\description{
Calculates number of crossing points, longest flat spot using the \code{feasts} package functionality and an additional time series feature - autocorrelation.
}
\examples{
pm_data <- get_wdi_data(indicator = "EN.ATM.PM25.MC.M3")
pm_temporal <- compute_temporal_features(pm_data)
}
