% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_trend_shape_features.R
\name{compute_trend_shape_features}
\alias{compute_trend_shape_features}
\title{Compute trend and shape features}
\usage{
compute_trend_shape_features(wdi_data, index = NULL)
}
\arguments{
\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}

\item{index}{An optional character string specifying the indicator code
Defaults to \code{NULL}}
}
\value{
A data frame with columns \code{country}, \code{trend_strength}, \code{linearity}, \code{curvature}, and \code{smoothness}.
}
\description{
Calculates trend strength, linearity, and curvature using the \code{feasts} and \code{fabletools} packages functionality.
}
\examples{
pm_trend_shape <- compute_trend_shape_features(pm_data)
}
