% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_groups.R
\docType{data}
\name{country_groups}
\alias{country_groups}
\title{Country with its grouping variables (region and income)}
\format{
A data frame with 3 variables
\describe{

\item{country}{The country name (character)}
\item{region}{The region grouping variable (character)}
\item{income}{The income grouping variable (character)}
}
}
\source{
World Development Indicator, using the WDI R package
}
\usage{
country_groups
}
\description{
This dataset contains the country names and their respective region and income groups
}
\examples{
data(country_groups)

head(country_groups)
}
\keyword{datasets}
