% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_data.R
\name{get_valid_data}
\alias{get_valid_data}
\title{Extract valid data from the WDI data
Reports countries with no data points as well as years for which no data are available.}
\usage{
get_valid_data(wdi_data, index = NULL)
}
\arguments{
\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}

\item{index}{An optional character string specifying the indicator code
Defaults to \code{NULL}}
}
\value{
A tibble with the valid data for the provided WDI indicator data set and a detailed report of missing entries.
}
\description{
Extract valid data from the WDI data
Reports countries with no data points as well as years for which no data are available.
}
\examples{
get_valid_data(pm_data)
}
