% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wdi_data.R
\name{get_wdi_data}
\alias{get_wdi_data}
\title{Download WDI data using the \code{WDI} R package}
\usage{
get_wdi_data(indicator)
}
\arguments{
\item{indicator}{A valid WDI indicator code}
}
\value{
An \code{.rds} file containing the data set for the specified indicator code.
}
\description{
Create and store the data for the specified indicator code in a folder called \code{wdi_data}.
}
\examples{
\dontrun{
 pm_data <- get_wdi_data(indicator = "EN.ATM.PM25.MC.M3")
 }
}
