% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_linkview.R
\name{plot_metric_linkview}
\alias{plot_metric_linkview}
\title{Plot of diagnostic metrics linked to data trajectories}
\usage{
plot_metric_linkview(
  wdi_data,
  index = NULL,
  metric_summary,
  metric_var,
  group_var = NULL
)
}
\arguments{
\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}

\item{index}{A character string specifying the indicator code
Defaults to \code{NULL}}

\item{metric_summary}{A data frame containing computed diagnostic metrics and the pre-defined grouping information,
generated by passing the output of any diagnostic metrics function to \code{\link{add_group_info}}}

\item{metric_var}{A vector of character strings specifying metric variable names in \code{metric_summary}}

\item{group_var}{A grouping variable in the WDI data set (e.g., "region" or "income")
Default to \code{NULL}
If \code{NULL}, both plots are ungrouped and if specified, they are grouped by the levels of the specified grouping variable}
}
\value{
An ungrouped or grouped interactive \code{girafe} object displaying the two panels, one with the scatterplot of two specified metrics and the other with the data trajectories.
}
\description{
Creates an interactive plot linking the scatterplot of two selected metrics with data trajectories.
The scatterplot showing the relationship between specified metrics are presented in one panel,
and the data trajectories are presented in another panel.
Hovering over a point in the scatterplot highlights the corresponding trajectory with the country name, and vice versa.
}
\examples{
pm_diagnostic_metrics <- compute_diagnostic_indices(pm_data, group_var = "region")
pm_diagnostic_metrics_group <- add_group_info(metric_summary = pm_diagnostic_metrics,pm_data)
plot_metric_linkview(pm_data, metric_summary = pm_diagnostic_metrics,
metric_var = c("linearity", "curvature"))
}
