% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Missingness plot of the indicator data}
\usage{
plot_missing(wdi_data, index = NULL, group_var)
}
\arguments{
\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}

\item{index}{An optional character string specifying the indicator code
Defaults to \code{NULL}}

\item{group_var}{A grouping variable in the WDI data set (e.g., "region" or "income")}
}
\value{
A plot that provides a structured overview of missing data and shows its distribution over time, across countries, and by the specified grouping variable.
}
\description{
Missingness plot of the indicator data
}
\examples{
plot_missing(pm_data, group_var = "region")
}
