% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dissimilarity.R
\name{compute_dissimilarity}
\alias{compute_dissimilarity}
\title{Compute dissimilarity between pair of countries
Calculate pairwise dissimilarities and convert the output to matrix.}
\usage{
compute_dissimilarity(wdi_data, index = NULL, metric = "euclidean")
}
\arguments{
\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}

\item{index}{An optional character string specifying the indicator code
Defaults to \code{NULL}}

\item{metric}{A character string specifying the dissimilarity metric to use
Defaults to \code{"euclidean"}}
}
\value{
A matrix of pairwise dissimilarities between countries.
}
\description{
Compute dissimilarity between pair of countries
Calculate pairwise dissimilarities and convert the output to matrix.
}
\examples{
pm_data <- get_wdi_data(indicator = "EN.ATM.PM25.MC.M3")
pm_diss_mat <- compute_dissimilarity(pm_data)
}
