% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variation.R
\name{compute_variation}
\alias{compute_variation}
\title{Compute variation features}
\usage{
compute_variation(
  wdi_data,
  diss_matrix = compute_dissimilarity(wdi_data),
  group_var
)
}
\arguments{
\item{wdi_data}{A data frame of the indicator data generated by \code{\link{get_wdi_data}}}

\item{diss_matrix}{An optional dissimilarity matrix generated by \code{\link{compute_dissimilarity}}}

\item{group_var}{A grouping variable in the WDI data set (e.g., "region" or "income")}
}
\value{
A data frame with columns \code{country}, \code{group}, \code{country_avg_dist}, \code{within_group_dist}, and \code{sil_width}.
}
\description{
Calculates average dissimilarities between countries, group-wise country dissimilarities, and silhouette widths.
}
\examples{
pm_data <- get_wdi_data(indicator = "EN.ATM.PM25.MC.M3")
pm_variation <- compute_variation(pm_data, group_var = "region")
}
